refactor: moved config logic into internals, added core types and config variables
